+++ NEWS.md
@@ -11,6 +11,7 @@
 - add support to save target data csv without change default name
 - add retention time alignment function for peaks lists
 - add getalign2 function to align mass to charge ratio, retention time for the same peak list, this function could be used to refine peaks list
+- fix color issues in plotrla and plotridge function
 
 # enviGCMS 0.6.6